requirements.txt
Flask
prometheus-client
python-keycloak
requests
opentelemetry-distro
opentelemetry-exporter-otlp-proto-http
